+++ src/periodic_trends/core.py
@@ -203,7 +203,7 @@ def plotter(
             color_list[element_index] = to_hex(color_scale[i])  # type: ignore
 
     if special_elements:
-        for k, v in elements["symbol"].iteritems():
+        for k, v in elements["symbol"].items():
             if v in special_elements:
                 color_list[k] = special_color